fix(layer): unbreak Cloudflare prerender and runtime bundle
Two issues surfaced in the Cloudflare Pages deploy of #27:

1. Crawler-discovered section index routes (e.g. /docs/getting-started)
   render via [...slug].vue even though no source markdown exists. The
   unconditional addPrerender call I added in PR #27 then registered
   /raw/docs/getting-started.md, which the raw handler 404s, failing the
   build with 'fatal: Page not found'. Gate the registration on
   page.value so we only enqueue raw paths for pages that actually have
   a source file.

2. The new sitemap.xml route imported inferSiteURL from utils/meta,
   which also exports getPackageJsonMetadata (node:fs/promises +
   node:path). That pulled node:* modules into the Worker bundle and
   broke the Cloudflare publish step ('No such module node:buffer').
   Split inferSiteURL out into a runtime-safe utils/site-url module and
   re-export it from utils/meta so existing build-time callers keep
   working.
